requirements.txt
requests
pandas
python-dotenv
yfinance
matplotlib
